fixed minor things

+++ Subproject2/0. Models/SearchResult.cs
@@ -7,7 +7,7 @@ namespace _0._Models
     public class SearchResult
     {
         public int Id { get; set; }
-        public int Rank { get; set; }
+        public double Rank { get; set; }
         public string Body { get; set; }
     }
 }
